+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/detection/annotation/Detection.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.linkedin.thirdeye.detection.annotations;
+package com.linkedin.thirdeye.detection.annotation;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.lang.annotation.ElementType;
